requirements.txt
dash==2.16.1
pandas==2.2.1
python-dotenv==1.0.1
scikit_learn==1.4.1.post1
spotipy==2.23.0
